Week 11-8-2023

+++ ResumableFunctions.Handler/Core/Abstraction/ICallProcessor.cs
@@ -4,7 +4,7 @@ namespace ResumableFunctions.Handler.Core.Abstraction
 {
     public interface ICallProcessor
     {
-        Task InitialProcessPushedCall(int pushedCallId, string methodUrn);
+        Task InitialProcessPushedCall(long pushedCallId, string methodUrn);
         Task ServiceProcessPushedCall(CallServiceImapction service);
     }
 }
